Basic untextured tileset
Importation of untextured tileset, a door that stays open for now. creation of a new scene.
within this new scene I've created a small demo level that can be used for further testing purposes, and possibly the new demo.
if you want to use it I would suggets copying the rooms to a different scene.